rename service UpdateWorkplaceAfterStaffChangesService --> VacanciesAndTurnoverService
