+++ frontend/src/app/features/workers/add-another-staff-record/add-another-staff-record.component.spec.ts
@@ -5,10 +5,10 @@ import { Router } from '@angular/router';
 import { EstablishmentService } from '@core/services/establishment.service';
 import {
   DoYouWantToAddOrDeleteAnswer,
-  UpdateWorkplaceAfterStaffChangesService,
-} from '@core/services/update-workplace-after-staff-changes.service';
+  VacanciesAndTurnoverService,
+} from '@core/services/vacancies-and-turnover.service';
 import { MockEstablishmentService } from '@core/test-utils/MockEstablishmentService';
-import { MockUpdateWorkplaceAfterStaffChangesService } from '@core/test-utils/MockUpdateWorkplaceAfterStaffChangesService';
+import { MockVacanciesAndTurnoverService } from '@core/test-utils/MockVacanciesAndTurnoverService';
 import { render } from '@testing-library/angular';
 import userEvent from '@testing-library/user-event';
 
@@ -27,8 +27,8 @@ describe('AddAnotherStaffRecordComponent', () => {
           useClass: MockEstablishmentService,
         },
         {
-          provide: UpdateWorkplaceAfterStaffChangesService,
-          useFactory: MockUpdateWorkplaceAfterStaffChangesService.factory({
+          provide: VacanciesAndTurnoverService,
+          useFactory: MockVacanciesAndTurnoverService.factory({
             resetVisitedAndSubmittedPages: resetVisitedAndSubmittedPagesSpy,
             ...overrides.updateWorkplaceService,
           }),
@@ -43,8 +43,8 @@ describe('AddAnotherStaffRecordComponent', () => {
     const navigateSpy = spyOn(router, 'navigate').and.returnValue(Promise.resolve(true));
 
     const updateWorkplaceAfterStaffChangesService = injector.inject(
-      UpdateWorkplaceAfterStaffChangesService,
-    ) as UpdateWorkplaceAfterStaffChangesService;
+      VacanciesAndTurnoverService,
+    ) as VacanciesAndTurnoverService;
     const doYouWantToAddOrDeleteAnswerSpy = spyOnProperty(
       updateWorkplaceAfterStaffChangesService,
       'doYouWantToAddOrDeleteAnswer',
