+++ frontend/src/app/features/workers/update-workplace-details-after-staff-changes/update-starters/update-starters.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { jobOptionsEnum } from '@core/model/establishment.model';
-import { WorkplaceUpdatePage } from '@core/services/update-workplace-after-staff-changes.service';
+import { WorkplaceUpdatePage } from '@core/services/vacancies-and-turnover.service';
 import { UpdateStartersLeaversVacanciesDirective } from '@shared/directives/update-starters-leavers-vacancies/update-starters-leavers-vacancies.directive';
 
 @Component({
